app/requirements.txt
paho-mqtt==2.1.0
flatbuffers==25.2.10 
numpy==2.2.3 
psycopg2==2.9.9
amqp==5.2.0
pycapnp==2.0.0
python-dotenv==1.1.0
